rgw/multisite: build a coroutine class to handle error repo logic

+++ src/rgw/rgw_data_sync.h
@@ -674,7 +674,7 @@ struct rgw_bucket_index_marker_info {
   bool syncstopped{false};
   uint64_t oldest_gen = 0;
   uint64_t latest_gen = 0;
-  std::vector<std::pair<int, int>> gen_numshards;
+  std::vector<std::pair<uint64_t, uint32_t>> gen_numshards;
 
   void decode_json(JSONObj *obj) {
     JSONDecoder::decode_json("bucket_ver", bucket_ver, obj);
